Xmlbf.Xeno

Documentation

element

Arguments

:: Node
A Node from Xeno.DOM.

-> Either String Node
A Node from Xmlbf.

Convert a Node from Xeno.DOM into an Element from Xmlbf.

nodes

Arguments

:: ByteString
Raw XML fragment.

-> Either String [Node]
Nodes from Xmlbf

Parses a given UTF8-encoded raw XML fragment into a, using the xeno Haskell library, so all of xeno's parsing quirks apply.

You can provide the output of this function as input to Xmlbf's runParser.

The given XML can contain more zero or more text or element nodes.

Surrounding whitespace is not stripped.